+++ modules/project/core/src/sap/commerce/toolset/project/PropertyService.kt
@@ -68,6 +68,7 @@ class PropertyService(private val project: Project) {
             var localPropsFile: PropertiesFile? = null
 
             val propertiesFiles = ArrayList<PropertiesFile>()
+            val platformHomeFiles = ArrayList<PropertiesFile>()
 
             // Ignore Order and production.properties for now as `developer.mode` should be set to true for development anyway
             FileTypeIndex.getFiles(PropertiesFileType.INSTANCE, scope)
@@ -79,6 +80,7 @@ class PropertyService(private val project: Project) {
                         HybrisConstants.ADVANCED_PROPERTIES_FILE -> advancedPropsFile = file
                         HybrisConstants.LOCAL_PROPERTIES_FILE -> localPropsFile = file
                         HybrisConstants.PROJECT_PROPERTIES_FILE -> propertiesFiles.add(file)
+                        HybrisConstants.PLATFORMHOME_PROPERTIES_FILE -> platformHomeFiles.add(file)
                     }
                 }
 
@@ -87,6 +89,7 @@ class PropertyService(private val project: Project) {
             envPropsFile?.let { propertiesFiles.add(0, it) }
 
             propertiesFiles.forEach { addPropertyFile(result, it) }
+            platformHomeFiles.forEach { addPropertyFile(result, it) }
 
             loadHybrisRuntimeProperties(result)
             loadHybrisOptionalConfigDir(result)
@@ -140,7 +143,7 @@ class PropertyService(private val project: Project) {
                 addEnvironmentProperties(properties)
                 properties
                     .filter { it.value.contains(nestedPropertyPrefix) }
-                    .forEach { replacePlaceholder(properties, it.key, HashSet<String>()) }
+                    .forEach { replacePlaceholder(properties, it.key, mutableSetOf()) }
                 return@let properties
             }
     }
@@ -176,8 +179,7 @@ class PropertyService(private val project: Project) {
             }
     }
 
-    private fun replacePlaceholder(result: LinkedHashMap<String, String>, key: String, visitedProperties: MutableSet<String>) {
-
+    private fun replacePlaceholder(result: MutableMap<String, String>, key: String, visitedProperties: MutableSet<String>): String {
         var lastIndex = 0
 
         val value = result[key] ?: ""
@@ -211,10 +213,20 @@ class PropertyService(private val project: Project) {
 
         }
         result[key] = replacedValue
+        return replacedValue
     }
 
     private fun loadHybrisOptionalConfigDir(result: MutableMap<String, IProperty>) = (System.getenv(HybrisConstants.ENV_HYBRIS_OPT_CONFIG_DIR)
         ?: result[HybrisConstants.PROPERTY_OPTIONAL_CONFIG_DIR]?.value)
+        ?.let { property ->
+            if (property.contains(nestedPropertyPrefix)) {
+                val properties = result
+                    .mapValues { it.value.value ?: "" }
+                    .toMutableMap()
+                properties["_____temp_property_opt_config_dir"] = property
+                replacePlaceholder(properties, "_____temp_property_opt_config_dir", mutableSetOf())
+            } else property
+        }
         ?.let { File(it) }
         ?.takeIf { it.isDirectory }
         ?.listFiles { _, name -> optionalPropertiesFilePattern.matcher(name).matches() }
@@ -245,11 +257,17 @@ class PropertyService(private val project: Project) {
     private fun createSearchScope(configModule: Module, platformModule: Module): GlobalSearchScope {
         val projectPropertiesScope = GlobalSearchScope.getScopeRestrictedByFileTypes(GlobalSearchScope.everythingScope(project), PropertiesFileType.INSTANCE)
             .filter { it.name == HybrisConstants.PROJECT_PROPERTIES_FILE }
+        val platformHomePropertiesScope = GlobalSearchScope.getScopeRestrictedByFileTypes(GlobalSearchScope.everythingScope(project), PropertiesFileType.INSTANCE)
+            .filter { it.name == HybrisConstants.PLATFORMHOME_PROPERTIES_FILE }
         val envPropertiesScope = platformModule.moduleContentScope.filter { it.name == HybrisConstants.ENV_PROPERTIES_FILE }
         val advancedPropertiesScope = platformModule.moduleContentScope.filter { it.name == HybrisConstants.ADVANCED_PROPERTIES_FILE }
         val localPropertiesScope = configModule.moduleContentScope.filter { it.name == HybrisConstants.LOCAL_PROPERTIES_FILE }
 
-        return envPropertiesScope.or(advancedPropertiesScope).or(localPropertiesScope).or(projectPropertiesScope)
+        return envPropertiesScope
+            .or(advancedPropertiesScope)
+            .or(platformHomePropertiesScope)
+            .or(localPropertiesScope)
+            .or(projectPropertiesScope)
     }
 
     private fun obtainConfigModule() = ModuleManager.getInstance(project)